Refactor EventDispatcher usage and clean up includes
Updated Application to use SharedPtr for EventDispatcher.
Removed unnecessary includes across multiple files,
replaced logging method, and improved include order
for better organization. Adjusted constructors and
function signatures to maintain consistency with
SharedPtr usage. Minor formatting changes were also made.
