+++ Lithe/include/Lithe/Core/Clock.h
