+++ Lithe/include/Lithe/Lithe.h
@@ -1,13 +1,26 @@
 #pragma once
 
-#include "Lithe/Core/Application.h"
-#include "Lithe/Core/Window.h"
-#include "Lithe/Core/Renderer.h"
+#include "Export.h"
+
+#include <Utils/Utils.h>
+
 #include "Lithe/Core/Log.h"
 
 #include "Lithe/Events/Event.h"
+#include "Lithe/Events/WindowEvents.h"
+#include "Lithe/Events/KeyEvents.h"
+#include "Lithe/Events/MouseEvents.h"
 #include "Lithe/Events/EventDispatcher.h"
 #include "Lithe/Events/EventBus.h"
 #include "Lithe/Events/Input.h"
+
+#include "Lithe/Scene/Scene.h"
+#include "Lithe/Scene/Camera.h"
+#include "Lithe/Scene/Entity.h"
+#include "Lithe/Scene/Components.h"
+
+#include "Lithe/Core/Window.h"
+#include "Lithe/Core/Renderer.h"
+#include "Lithe/Core/Application.h"
 
 #include "Lithe/Utils/Utils.h"
