+++ Lithe/include/Lithe/Scene/Camera.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Lithe/Core/Clock.h"
+#include "Clock.h"
 #include <glm/glm.hpp>
 
 namespace Lithe {
