+++ Lithe/src/Core/EntryPoint.cpp
@@ -7,8 +7,8 @@ LITHE_EXPORT void Lithe::start(Lithe::ApplicationCreateFunc create) {
 	using namespace Lithe;
 	Logger::init("[%Y-%m-%d %H:%M:%S.%e] [%^%l%$] %v");
 
-	EventDispatcher dispatcher;
-	EventBus bus(dispatcher);
+	auto dispatcher = makeShared<EventDispatcher>();
+	EventBus bus(*dispatcher);
 
 
 	Application* application = create(dispatcher);
