+++ Lithe/src/Scene/OrthographicCamera.cpp
@@ -1,11 +1,9 @@
 #include "pch.h"
 #include "OrthographicCamera.h"
 
-#include "Lithe/Events/Event.h"
-#include "Lithe/Events/Input.h"
+#include "Input.h"
 
 #include <glm/gtc/matrix_transform.hpp>
-
 
 namespace Lithe {
