+++ Sandbox/Sandbox.cpp
@@ -4,7 +4,7 @@ class Sandbox: public Lithe::Application {
 
 	public:
 
-		Sandbox(Lithe::EventDispatcher& dispatcher): Application(dispatcher) {
+		Sandbox(Lithe::SharedPtr<Lithe::EventDispatcher> dispatcher): Application(dispatcher) {
 			std::cout << "Sandbox created\n";
 		}
 
@@ -14,7 +14,7 @@ class Sandbox: public Lithe::Application {
 	
 };
 
-Lithe::Application* create(Lithe::EventDispatcher& dispatcher) {
+Lithe::Application* create(Lithe::SharedPtr<Lithe::EventDispatcher> dispatcher) {
 	return new Sandbox(dispatcher);
 }
